[bcr-wdc-quote-service]: move RANDOM and publics() to bcr-wdc-keys
